Vorschau eingebaut

+++ ulicms/content/modules/text_rotator/templates/list.php
@@ -1,26 +1,38 @@
 <?php
 
 use UliCMS\HTML\Input;
+use UliCMS\Security\PermissionChecker;
 
 $rotatingTexts = RotatingText::getAll();
+$permissionChecker = new PermissionChecker(get_user_id());
 ?>
-<div class="form-group">
-    <a href="<?php esc(ModuleHelper::buildActionURL("text_rotator_create")); ?>" class="btn btn-primary">
-        <i class="fa fa-plus"></i> <?php translate("new"); ?></a>
-</div>
-<div class="scrollable">
+<?php
+if ($permissionChecker->hasPermission("text_rotator_edit")) {
+    ?>
+    <div class="form-group">
+        <a href="<?php esc(ModuleHelper::buildActionURL("text_rotator_create")); ?>" clas="btn btn-primary">
+            <i class="fa fa-plus"></i> <?php translate("new");
+    ?></a>
+    </div>
+<?php } ?>
+<div class="scroll">
     <table class="tablesorter">
         <thead>
             <tr>
                 <th><?php translate("id"); ?></th>
                 <th><?php translate("words"); ?></th>
                 <th class="no-sort"><?php translate("shortcode"); ?></th>
-                <th class="no-sort text-center">
-                    <?php translate("edit") ?>
-                </th>
-                <th class="no-sort text-center">
-                    <?php translate("delete") ?>
-                </th>
+                <?php
+                if ($permissionChecker->hasPermission("text_rotator_edit")) {
+                    ?>
+                    <th class="no-sort text-center">
+                        <?php translate("edit")
+                        ?>
+                    </th>
+                    <th class="no-sort text-center">
+                        <?php translate("delete") ?>
+                    </th>
+                <?php } ?>
             </tr>
         </thead>
         <tbody>
@@ -37,18 +49,22 @@ $rotatingTexts = RotatingText::getAll();
                                     "class" => "select-on-click"));
                         ?>
                     </td>
-                    <td class="text-center"><a
-                            href="<?php
-                            esc(ModuleHelper::buildActionURL("text_rotator_edit", "id={$text->getID()}"));
-                            ?>">
-                            <img class="mobile-big-image" src="gfx/edit.png"
-                                 alt="<?php translate("edit"); ?>"
-                                 title="<?php translate("edit"); ?>">
-                        </a></td>
-                    <td class="text-center">
-                        <?php echo ModuleHelper::deleteButton(ModuleHelper::buildMethodCallUrl(TextRotatorController::class, "delete", "id={$text->getID()}")); ?>
+                    <?php
+                    if ($permissionChecker->hasPermission("text_rotator_edit")) {
+                        ?>
+                        <td class="text-center"><a
+                                href="<?php
+                                esc(ModuleHelper::buildActionURL("text_rotator_edit", "id={$text->getID()}"));
+                                ?>">
+                                <img class="mobile-big-image" src="gfx/edit.png"
+                                     alt="<?php translate("edit"); ?>"
+                                     title="<?php translate("edit"); ?>">
+                            </a></td>
+                        <td class="text-center">
+                            <?php echo ModuleHelper::deleteButton(ModuleHelper::buildMethodCallUrl(TextRotatorController::class, "delete", "id={$text->getID()}")); ?>
 
-                    </td>
+                        </td>
+                    <?php } ?>
                 </tr>
                 <?php
             }
